Extracted import code from menu to dedicated class
