+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Import/VpxImportEngine.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5d08ade357dc1dc4d910791dff0665ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
