+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Import/VpxMenuImporter.cs
@@ -2,21 +2,13 @@
 // ReSharper disable UnusedType.Global
 // ReSharper disable UnusedMember.Global
 
-using System.IO;
-using NLog;
 using UnityEditor;
 using UnityEngine;
-using VisualPinball.Unity.Editor.Utils;
-using VisualPinball.Unity.Import;
-using VisualPinball.Unity.Import.Job;
-using Logger = NLog.Logger;
 
 namespace VisualPinball.Unity.Editor.Import
 {
 	public static class VpxMenuImporter
 	{
-
-		private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
 
 		[MenuItem("Visual Pinball/Import VPX", false, 10)]
 		public static void ImportVpxEditorMemory(MenuCommand menuCommand)
@@ -41,32 +33,8 @@ namespace VisualPinball.Unity.Editor.Import
 				return;
 			}
 
-			var rootGameObj = ImportVpx(vpxPath);
-
-			// if an object was selected in the editor, make it its parent
-			GameObjectUtility.SetParentAndAlign(rootGameObj, menuCommand.context as GameObject);
-
-			// register undo system
-			Undo.RegisterCreatedObjectUndo(rootGameObj, "Import VPX table file");
-
-			// select imported object
-			Selection.activeObject = rootGameObj;
-
-			Logger.Info("[VpxImporter] Imported!");
-		}
-
-		private static GameObject ImportVpx(string path) {
-
-			// create root object
-			var rootGameObj = new GameObject();
-			var importer = rootGameObj.AddComponent<VpxImporter>();
-
-			// load table
-			var table = TableLoader.LoadTable(path);
-
-			importer.Import(Path.GetFileName(path), table);
-
-			return rootGameObj;
+			// perform import
+			VpxImportEngine.Import(vpxPath, menuCommand.context as GameObject);
 		}
 	}
 }
